+++ server/src/routes/api/v2/relay.js
@@ -1,12 +1,12 @@
 const router = require('express').Router()
-const { agenda } = require('@services/agenda')
+const taskManager = require('@services/taskManager')
 const auth = require('@routes/auth')
 
 router.post('/', auth.required, async (req, res) => {
   const { appName, identifier } = req.user
   try {
-    const job = await agenda.now('relay', { ...req.body, identifier, appName })
-    return res.json({ job: job.attrs })
+    const job = await taskManager.now('relay', { ...req.body, identifier, appName })
+    return res.json({ job })
   } catch (err) {
     return res.status(400).send({ error: err.message })
   }
@@ -20,8 +20,8 @@ router.post('/multi', auth.required, async (req, res) => {
       return res.status(400).send({ error: `No items in body` })
     }
     const firstItem = items.shift()
-    const job = await agenda.now('relay', { ...firstItem, identifier, appName, nextRelays: items })
-    return res.json({ job: job.attrs })
+    const job = await taskManager.now('relay', { ...firstItem, identifier, appName, nextRelays: items })
+    return res.json({ job })
   } catch (err) {
     console.error(err)
     return res.status(400).send({ error: err.message })
