+++ server/src/routes/api/v2/wallets/index.js
@@ -1,6 +1,7 @@
 const config = require('config')
 const homeAddresses = config.get('network.home.addresses')
 const router = require('express').Router()
+const taskManager = require('@services/taskManager')
 const { agenda } = require('@services/agenda')
 const { web3 } = require('@services/web3/home')
 const auth = require('@routes/auth')
@@ -28,8 +29,8 @@ router.post('/', auth.required, async (req, res, next) => {
   const transferOwnerWallet = await UserWallet.findOne({ phoneNumber, accountAddress: config.get('network.home.addresses.MultiSigWallet') })
   if (transferOwnerWallet) {
     console.log(`User ${phoneNumber} already has wallet account: ${transferOwnerWallet.walletAddress} owned by MultiSig - need to setOwner`)
-    const job = await agenda.now('setWalletOwner', { walletAddress: transferOwnerWallet.walletAddress, newOwner: accountAddress, correlationId })
-    return res.json({ job: job.attrs })
+    const job = await taskManager.now('setWalletOwner', { walletAddress: transferOwnerWallet.walletAddress, newOwner: accountAddress, correlationId })
+    return res.json({ job: job })
   } else {
     let userWallet = await UserWallet.findOne({ phoneNumber, accountAddress, appName })
     if (userWallet) {
@@ -51,8 +52,8 @@ router.post('/', auth.required, async (req, res, next) => {
         appName,
         ip: req.clientIp
       }).save()
-      const job = await agenda.now('createWallet', { owner: accountAddress, correlationId, _id: userWallet._id })
-      return res.json({ job: job.attrs })
+      const job = await taskManager.now('createWallet', { owner: accountAddress, correlationId, _id: userWallet._id })
+      return res.json({ job: job })
     }
   }
 })
@@ -207,9 +208,9 @@ router.post('/invite/:phoneNumber', auth.required, async (req, res, next) => {
     appName
   }).save()
 
-  const job = await agenda.now('createWallet', { owner, communityAddress, phoneNumber: invitedPhoneNumber, name, amount, symbol, bonusInfo, correlationId, _id: userWallet._id, appName })
+  const job = await taskManager.now('createWallet', { owner, communityAddress, phoneNumber: invitedPhoneNumber, name, amount, symbol, bonusInfo, correlationId, _id: userWallet._id, appName })
 
-  return res.json({ job: job.attrs })
+  return res.json({ job })
 })
 
 /**
@@ -250,8 +251,8 @@ router.post('/backup', auth.required, async (req, res, next) => {
       bonusId: phoneNumber
     }
 
-    const job = await agenda.now('bonus', { communityAddress, bonusInfo, correlationId })
-    return res.json({ job: job.attrs })
+    const job = await taskManager.now('bonus', { communityAddress, bonusInfo, correlationId })
+    return res.json({ job: job })
   }
 
   return res.json({ response: 'ok' })
