--- server/src/routes/api/v3/index.js
@@ -1,8 +1,0 @@
-const router = require('express').Router()
-
-router.use('/relay', require('./relay'))
-router.use('/wallets', require('./wallets'))
-router.use('/jobs', require('./jobs'))
-router.use('/fund', require('./fund'))
-
-module.exports = router
